bt-662: * code cleanup
