+++ backend/src/bundles/lms-data/constants/constants.ts
@@ -1,0 +1,1 @@
+export { defaultShownBadges } from './default-shown-badges.js';
