+++ backend/src/bundles/lms-data/enums/enums.ts
@@ -1,0 +1,1 @@
+export { BsaBadgesTypeEnum } from 'shared/build/index.js';
