+++ backend/src/bundles/lms-data/helpers/helpers.ts
@@ -1,1 +1,3 @@
+export { getRandomScore } from './get-random-score.js';
+export { isNumber } from './is-number.js';
 export { makeLMSDataResponse, parseLMSServerData } from 'shared/build/index.js';
