+++ backend/src/bundles/lms-data/helpers/is-number.ts
@@ -1,0 +1,5 @@
+const isNumber = (value: unknown): value is number => {
+    return typeof value === 'number' && !Number.isNaN(value);
+};
+
+export { isNumber };
