+++ backend/src/bundles/lms-data/mappers/mappers.ts
@@ -1,0 +1,1 @@
+export { badgeTypeToLMSProperty } from './badge-type-to-lms-property.js';
