+++ backend/src/bundles/talent-badges/talent-badge.entity.ts
@@ -1,6 +1,5 @@
 import { type Entity } from '~/common/types/types.js';
 
-import { type BSABadgeEntity } from '../bsa-badges/bsa-badges.entity.js';
 import { type BadgesItem } from '../lms-data/types/types.js';
 import { type TalentBadgeCreateDto } from './types/types.js';
 
@@ -19,7 +18,7 @@ class TalentBadgeEntity implements Entity {
 
     private 'userDetailsId': string;
 
-    private 'badge': BSABadgeEntity | null;
+    private 'badge': BadgesItem | null;
 
     private constructor({
         id,
@@ -78,7 +77,7 @@ class TalentBadgeEntity implements Entity {
         badgeId: string;
         isShown: boolean;
         userDetailsId: string;
-        badge: BSABadgeEntity | null;
+        badge: BadgesItem | null;
     }): TalentBadgeEntity {
         return new TalentBadgeEntity({
             id: null,
@@ -110,7 +109,7 @@ class TalentBadgeEntity implements Entity {
             badgeId: this.badgeId,
             isShown: this.isShown,
             userDetailsId: this.userDetailsId,
-            badge: this.badge?.toObject() ?? null,
+            badge: this.badge ?? null,
         };
     }
